distributing func in files according to their behavior
