+++ internal/controller/atlasipaccesslist/atlasipaccesslist_controller_test.go
@@ -2,11 +2,8 @@ package atlasipaccesslist
 
 import (
 	"context"
-	"errors"
 	"testing"
 
-	"github.com/google/go-cmp/cmp"
-	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
@@ -26,7 +23,6 @@ import (
 	akov2 "github.com/mongodb/mongodb-atlas-kubernetes/v2/api/v1"
 	"github.com/mongodb/mongodb-atlas-kubernetes/v2/api/v1/common"
 	"github.com/mongodb/mongodb-atlas-kubernetes/v2/internal/controller/atlas"
-	"github.com/mongodb/mongodb-atlas-kubernetes/v2/internal/controller/customresource"
 	"github.com/mongodb/mongodb-atlas-kubernetes/v2/internal/controller/reconciler"
 	"github.com/mongodb/mongodb-atlas-kubernetes/v2/internal/controller/workflow"
 	atlasmock "github.com/mongodb/mongodb-atlas-kubernetes/v2/internal/mocks/atlas"
@@ -133,294 +129,6 @@ func TestReconcile(t *testing.T) {
 			require.NoError(t, err)
 
 			assert.Equal(t, tt.expectedResult, result)
-		})
-	}
-}
-
-func TestEnsureCustomResource(t *testing.T) {
-	tests := map[string]struct {
-		ipAccessList       akov2.AtlasIPAccessList
-		provider           atlas.Provider
-		expectedResult     ctrl.Result
-		expectedFinalizers []string
-		expectedConditions []api.Condition
-	}{
-		"should skip reconciliation": {
-			ipAccessList: akov2.AtlasIPAccessList{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "ip-access-list",
-					Namespace: "default",
-					Annotations: map[string]string{
-						customresource.ReconciliationPolicyAnnotation: customresource.ReconciliationPolicySkip,
-					},
-					Finalizers: []string{customresource.FinalizerLabel},
-				},
-				Spec: akov2.AtlasIPAccessListSpec{
-					Entries: []akov2.IPAccessEntry{
-						{
-							CIDRBlock: "192.168.0.0/24",
-						},
-					},
-				},
-			},
-			expectedResult:     ctrl.Result{},
-			expectedFinalizers: []string{customresource.FinalizerLabel},
-		},
-		"should fail to validate resource": {
-			ipAccessList: akov2.AtlasIPAccessList{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "ip-access-list",
-					Namespace: "default",
-					Labels: map[string]string{
-						customresource.ResourceVersion: "wrong",
-					},
-				},
-				Spec: akov2.AtlasIPAccessListSpec{
-					Entries: []akov2.IPAccessEntry{
-						{
-							CIDRBlock: "192.168.0.0/24",
-						},
-					},
-				},
-			},
-			expectedResult: ctrl.Result{RequeueAfter: workflow.DefaultRetry},
-			expectedConditions: []api.Condition{
-				api.FalseCondition(api.ReadyType),
-				api.FalseCondition(api.ResourceVersionStatus).
-					WithReason(string(workflow.AtlasResourceVersionIsInvalid)).
-					WithMessageRegexp("wrong is not a valid semver version for label mongodb.com/atlas-resource-version"),
-			},
-		},
-		"should fail when not supported": {
-			ipAccessList: akov2.AtlasIPAccessList{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "ip-access-list",
-					Namespace: "default",
-				},
-				Spec: akov2.AtlasIPAccessListSpec{
-					Entries: []akov2.IPAccessEntry{
-						{
-							CIDRBlock: "192.168.0.0/24",
-						},
-					},
-				},
-			},
-			provider: &atlasmock.TestProvider{
-				IsSupportedFunc: func() bool {
-					return false
-				},
-			},
-			expectedResult: ctrl.Result{},
-			expectedConditions: []api.Condition{
-				api.FalseCondition(api.ReadyType).
-					WithReason(string(workflow.AtlasGovUnsupported)).
-					WithMessageRegexp("the AtlasIPAccessList is not supported by Atlas for government"),
-				api.TrueCondition(api.ResourceVersionStatus),
-			},
-		},
-		"should fail to resolve credentials": {
-			ipAccessList: akov2.AtlasIPAccessList{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "ip-access-list",
-					Namespace: "default",
-				},
-				Spec: akov2.AtlasIPAccessListSpec{
-					ProjectDualReference: akov2.ProjectDualReference{
-						ProjectRef: &common.ResourceRefNamespaced{
-							Name: "my-no-existing-project",
-						},
-					},
-					Entries: []akov2.IPAccessEntry{
-						{
-							CIDRBlock: "192.168.0.0/24",
-						},
-					},
-				},
-			},
-			provider: &atlasmock.TestProvider{
-				IsSupportedFunc: func() bool {
-					return true
-				},
-			},
-			expectedResult: ctrl.Result{RequeueAfter: workflow.DefaultRetry},
-			expectedConditions: []api.Condition{
-				api.FalseCondition(api.ReadyType).
-					WithReason(string(workflow.AtlasAPIAccessNotConfigured)).
-					WithMessageRegexp("can not fetch AtlasProject: atlasprojects.atlas.mongodb.com \"my-no-existing-project\" not found"),
-				api.TrueCondition(api.ResourceVersionStatus),
-			},
-		},
-		"should fail to create sdk": {
-			ipAccessList: akov2.AtlasIPAccessList{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "ip-access-list",
-					Namespace: "default",
-				},
-				Spec: akov2.AtlasIPAccessListSpec{
-					ProjectDualReference: akov2.ProjectDualReference{
-						ConnectionSecret: &api.LocalObjectReference{
-							Name: "my-secret",
-						},
-					},
-					Entries: []akov2.IPAccessEntry{
-						{
-							CIDRBlock: "192.168.0.0/24",
-						},
-					},
-				},
-			},
-			provider: &atlasmock.TestProvider{
-				IsSupportedFunc: func() bool {
-					return true
-				},
-				SdkClientFunc: func(secretRef *client.ObjectKey, log *zap.SugaredLogger) (*admin.APIClient, string, error) {
-					return nil, "", errors.New("failed to create sdk")
-				},
-			},
-			expectedResult: ctrl.Result{RequeueAfter: workflow.DefaultRetry},
-			expectedConditions: []api.Condition{
-				api.FalseCondition(api.ReadyType).
-					WithReason(string(workflow.AtlasAPIAccessNotConfigured)).
-					WithMessageRegexp("failed to create sdk"),
-				api.TrueCondition(api.ResourceVersionStatus),
-			},
-		},
-		"should fail to resolve project": {
-			ipAccessList: akov2.AtlasIPAccessList{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "ip-access-list",
-					Namespace: "default",
-				},
-				Spec: akov2.AtlasIPAccessListSpec{
-					ProjectDualReference: akov2.ProjectDualReference{
-						ConnectionSecret: &api.LocalObjectReference{
-							Name: "my-secret",
-						},
-						ProjectRef: &common.ResourceRefNamespaced{
-							Name: "my-no-existing-project",
-						},
-					},
-					Entries: []akov2.IPAccessEntry{
-						{
-							CIDRBlock: "192.168.0.0/24",
-						},
-					},
-				},
-			},
-			provider: &atlasmock.TestProvider{
-				IsSupportedFunc: func() bool {
-					return true
-				},
-				SdkClientFunc: func(secretRef *client.ObjectKey, log *zap.SugaredLogger) (*admin.APIClient, string, error) {
-					return &admin.APIClient{}, "", nil
-				},
-			},
-			expectedResult: ctrl.Result{RequeueAfter: workflow.DefaultRetry},
-			expectedConditions: []api.Condition{
-				api.FalseCondition(api.ReadyType).
-					WithReason(string(workflow.AtlasAPIAccessNotConfigured)).
-					WithMessageRegexp("failed to query Kubernetes: failed to get Project from Kubernetes: can not fetch AtlasProject: atlasprojects.atlas.mongodb.com \"my-no-existing-project\" not found"),
-				api.TrueCondition(api.ResourceVersionStatus),
-			},
-		},
-		"should handle ip access list": {
-			ipAccessList: akov2.AtlasIPAccessList{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "ip-access-list",
-					Namespace: "default",
-				},
-				Spec: akov2.AtlasIPAccessListSpec{
-					ProjectDualReference: akov2.ProjectDualReference{
-						ConnectionSecret: &api.LocalObjectReference{
-							Name: "my-secret",
-						},
-						ProjectRef: &common.ResourceRefNamespaced{
-							Name: "my-project",
-						},
-					},
-					Entries: []akov2.IPAccessEntry{
-						{
-							CIDRBlock: "192.168.0.0/24",
-						},
-					},
-				},
-			},
-			//nolint:dupl
-			provider: &atlasmock.TestProvider{
-				IsSupportedFunc: func() bool {
-					return true
-				},
-				SdkClientFunc: func(secretRef *client.ObjectKey, log *zap.SugaredLogger) (*admin.APIClient, string, error) {
-					ialAPI := mockadmin.NewProjectIPAccessListApi(t)
-					ialAPI.EXPECT().ListProjectIpAccessLists(mock.Anything, "").
-						Return(admin.ListProjectIpAccessListsApiRequest{ApiService: ialAPI})
-					ialAPI.EXPECT().ListProjectIpAccessListsExecute(mock.AnythingOfType("admin.ListProjectIpAccessListsApiRequest")).
-						Return(
-							&admin.PaginatedNetworkAccess{
-								Results: &[]admin.NetworkPermissionEntry{
-									{
-										CidrBlock: pointer.MakePtr("192.168.0.0/24"),
-									},
-								},
-							},
-							nil,
-							nil,
-						)
-					ialAPI.EXPECT().GetProjectIpAccessListStatus(mock.Anything, "", "192.168.0.0/24").
-						Return(admin.GetProjectIpAccessListStatusApiRequest{ApiService: ialAPI})
-					ialAPI.EXPECT().GetProjectIpAccessListStatusExecute(mock.AnythingOfType("admin.GetProjectIpAccessListStatusApiRequest")).
-						Return(
-							&admin.NetworkPermissionEntryStatus{STATUS: "ACTIVE"},
-							nil,
-							nil,
-						)
-					return &admin.APIClient{ProjectIPAccessListApi: ialAPI}, "", nil
-				},
-			},
-			expectedResult:     ctrl.Result{},
-			expectedFinalizers: []string{customresource.FinalizerLabel},
-			expectedConditions: []api.Condition{
-				api.TrueCondition(api.ReadyType),
-				api.TrueCondition(api.ResourceVersionStatus),
-				api.TrueCondition(api.IPAccessListReady),
-			},
-		},
-	}
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			project := &akov2.AtlasProject{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "my-project",
-					Namespace: "default",
-				},
-			}
-			testScheme := runtime.NewScheme()
-			require.NoError(t, akov2.AddToScheme(testScheme))
-			k8sClient := fake.NewClientBuilder().
-				WithScheme(testScheme).
-				WithObjects(project, &tt.ipAccessList).
-				WithStatusSubresource(&tt.ipAccessList).
-				Build()
-			logger := zaptest.NewLogger(t).Sugar()
-			ctx := &workflow.Context{
-				Context: context.Background(),
-				Log:     logger,
-			}
-			r := &AtlasIPAccessListReconciler{
-				AtlasReconciler: reconciler.AtlasReconciler{
-					Client: k8sClient,
-					Log:    logger,
-				},
-				AtlasProvider: tt.provider,
-				EventRecorder: record.NewFakeRecorder(10),
-			}
-			result := r.ensureCustomResource(ctx.Context, &tt.ipAccessList)
-
-			ipAccessList := &akov2.AtlasIPAccessList{}
-			require.NoError(t, k8sClient.Get(ctx.Context, client.ObjectKeyFromObject(&tt.ipAccessList), ipAccessList))
-			assert.Equal(t, tt.expectedResult, result)
-			assert.Equal(t, tt.expectedFinalizers, ipAccessList.GetFinalizers())
-			assert.True(t, cmp.Equal(tt.expectedConditions, ipAccessList.Status.GetConditions(), cmpopts.IgnoreFields(api.Condition{}, "LastTransitionTime")))
 		})
 	}
 }
